Rebase CE branch onto CE/EE merge base before diff
If CE master is ahead of EE master, and the CE branch
was forked from CE master, then the CE branch would
contain a lot of changes which aren't presented in EE master,
therefore they're not presented in EE branch, either.

In this case, we need to remove the commits from CE master
which aren't merged into EE yet. So we want to rebase CE branch
onto CE/EE merge base, with ce_fetch_base and ce_fetch_head,
where ce_fetch_base is the merge base of CE master and CE branch,
and ce_fetch_head is the original HEAD of CE branch.

This should make sure there's no extra commits involved.

The only way it could fail would be conflicts during rebase,
which ideally should not happen at all. This doesn't mean
it's impossible to happen, but given the reason we make the CE branch,
it should only contain minimum changes from EE to CE, which shouldn't
conflict.

Let's try this and see and investigate if conflicts happen.