Reduce default wait interval
Because if reload is true, then we spend a lot of time waiting for page load and yielding (to has_css? etc, which already waits for the default Capybara time) anyway. And if reload is false, there’s no reason not to check what’s on the page at all times.

There doesn’t seem to be a downside.