Pass the template used to create Project between services
Before enqueuing the job to import a project we need to know if it was
created from a template in order to execute some tasks. We have the `template_name`
param but it's already being used to invoke the `Projects::CreateFromTemplateService`
service and it's logically being deleted from `params` otherwise we will
generate an infinite recursive call. That's why we need to use a
different param name to persist the template name.