Add a unique constraint on the project_features.project_id column
This migration prunes duplicate rows in the `project_features` table and leaves
the row with the highest ID.  Since the behavior was indeterministic before
and depended on which row the database decided to use before, this change at
least makes the permissions consistent. For example, in some cases, the Wiki
may have been disabled but enabled in another entry.

Fixes a significant part of https://gitlab.com/gitlab-com/migration/issues/408.
We found that we were overcounting Wikis because of these duplicates.