Update CE branch whenever EE is ahead of CE branch
This would happen if the corresponding CE branch is behind
CE master, and the CE merge base is ahead of CE branch.

When this happened, we could try to perform a merge to
CE merge base so that we keep up-to-date with the EE branch.

If there's a conflict, the author should resolve it from
the CE branch first.