edit GitLab license info in regards to CC licensing
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit moves the reference to Creative Commons licensing from
the LICENSE file to the README, which allows Licensee to successfully
identify the license type of GitLab CE as MIT.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>