Add service discovery for the DB load balancer
The service discovery mechanism can be used to automatically update the
list of database secondaries. This is done by periodically checking a
DNS record, updating the load balancer whenever its list of hosts
differs from the DNS record.

Old connections are terminated after the new hosts have been added to
the load balancer. A timeout mechanism is put in place to ensure this
completes in a fixed amount of time. Given requests check their
connections back in at the end of a request, the usual time this takes
should be in the order of a few seconds.