[Rails5] Explicitly set request.format for blob_controller
In Rails 4.2 if params[:format] is empty, Rails set it to :html
But since Rails 5.0 the framework now looks for an extension.
E.g. for `blob/master/CHANGELOG.md` in Rails 4 the format would be
`:html`, but in Rails 5 on it'd be `:md`

This commit explicitly sets the `:html` format for all requests unless
`:format` is set by a client e.g. by JS for XHR requests.