[Rails5] Fix ActionCable '/cable' mountpoint conflict
Since Rails 5.0 the new framework has been added. It's called ActionCable.
It brings WebSockets support to rails applications.
By default the mountpoint of WebSocket requests is `/cable` [1].

GitLab allows using top level names as namespaces or usernames.
For example, `gitlab.com/cable` at this moment leads to a user with the
nickname `cable`.

This commit changes ActionCable's mountpoint to a reserved top level
word `-`.
This is just a stub and should not be be used in real work.

Please set correct mountpoints for each environments when configuring
ActionCable for real using.

[1]: https://github.com/rails/rails/blob/5-0-stable/actioncable/lib/action_cable.rb#L38