Exclude Geo DB Yaml on CE too
For Geo, a file is created for configuration of the database. This file
shouldn't appear in CE, but if you've setup your GDK with a remote
pointing to CE and one to EE, this file might be around when switching
branches.

This will mark the workspace dirty, which annoys me more than it should.

[ci skip]