Add draft comments for MergeRequests
Implements a new model, DraftNote, which is related to MergeRequests
only and enables to comment on a review in batches. The DraftNotes are
only visible to the author until they're published at which point they
become regular Notes.

This also includes a DraftsController and two related services:
DraftNotes::CreateService and DraftNotes::PublishService

A DraftNote behaves almost identically to a Note, except has no cached
markdown fields and needs to process quick action changes every time it
is displayed.

NOTE: THIS IS ONLY THE BACKEND CHANGES. NO FRONTEND INCLUDED