Start performance bar eagerly
On the issues page (for instance), we have an `Issue` class. The constructor for
that class makes some calls to get things like related branches using Axios.
However, those calls were happening before the performance bar app was mounted,
and so the interceptor wasn't configured - meaning we missed out on the
performance data for those calls.

This changes the performance bar app to be loaded eagerly instead of lazily,
which solves that problem - related branches calls will now show up in the
performance bar.

The downside is that now the performance bar app is always imported, even when
it's not used. However, it appears to be part of the main app bundle anyway, so
this shouldn't affect the size of our assets.