Unify app/views/shared/issuable/_search_bar.html.haml
This is a bit tricky, so here we tried to port some codes to CE
if it doesn't hurt. If we tried to keep the behaviour completely
the same, the code would be pretty complicated without a good reason.

This also tries to reduce the changes by not moving a block into a
partial, which would force indentation and a lot of changes.