Remove trailing comma from weight system notes
We added a comma to the end of these to disambiguate between the value and the
timestamp after it. In the following release, we added a dot between the note
and the timestamp, so the comma was redundant.

This uses a background migration to find notes that are:

1. System notes.
2. That are about weight.
3. That end in a comma.

And remove the comma and clear the cached HTML. There are about 15,000 of these
on GitLab.com.