Fix handling of annotated tags when Gitaly is not in use
Attempting to view an annotated tag in the TreeController would result
in `NoMethodError: undefined method 'tree'` when Rugged was in use.
`Blob#find_by_rugged` assumes that the ref is a true. Using
the commit ID ensures that the right ref is being used.

Note that in 11.1, `Blob#find` no longer uses Rugged, so this is only
a bug in 11.0.

Closes gitlab-org/gitlab-ce#47797