Add existence checks around todos migration to ensure it is idempotent
Since this migration disables transactions, it's possible that the columns
and indexes are created but the migration is aborted in
the  `change_column_null` column. Check the existence of the
columns/foreign keys/indexes before attempting to add them to prevent
migration failures.

Relates to https://gitlab.com/gitlab-org/release/tasks/issues/312