Set AUTO_DEVOPS_DOMAIN default based on the ingress external_ip to enable zero-config domain name for the Auto Devops flow.
Allow the ingress default to be overriddable by the
`predefined_variables` of the project's `auto_devops`

Uses #deployment_platform which already handles all the logic to find
the cluster with `*` environment scope.

Create #cluster_ingress_domain as its own method so that it can be used in the frontend
- for help text for the autodevops domain input field and also to decide
if the warning message about `missing_auto_devops_domain` is shown or
not.

Add Changelog entry

DRY up the tests for checking which value for the domain is used

Allow the environment scope for the cluster to be passed in for EE
purposes