Remove direct disk access in RemoveDotGitFromUsernames
Prior to this change, the migration touched the disk path. This has been
removed.

Further, I believe it also fixes a bug. It seems that GitLab Shell
checks if the root path exists, when the path is duplicated. For
example, if the shard is located at /home/git, it did check if
`/home/git/home/git` existed.