Fix slow Markdown rendering
The sanitize transformers were being duplicated each time the Markdown
renderer was called, leading to expontential growth in rendering times.

The problem was that although HTML::Pipeline::SanitizationFilter.WHITELIST
is a frozen hash, the `:transformers` array can be modified. We need
to do deep copy of this to avoid adding duplicates.

Closes #49409