Require has_internal_id to pass a init block.
This is useful since the current scheme allows us to freely delete any
records in internal_ids. If the record is not there, it will get rebuild
using the init block.

We require passing in 'init for the time being and as long as its
not expensive to do so. This makes the scheme a bit more robust.

If at some point, this is expensive - we may choose to make this
optional while losing the ability to recalculate the value on the fly.

Closes #49609.