Add clarification points to using SCP style URIs
This addresses user-facing documentation related to repository
mirroring.

On the repositoring mirroring view of repository settings, inputting
either git@example.com:group/project.git or any variation of schemes:

* https://git@example.com:group/project.git
* http://git@example.com:group/project.git
* git://git@example.com:group/project.git
* ssh://git@example.com:group/project.git

Fails on URL validation.

This commit adds notes about the required format to the doc and view for
repository mirroring. This commit does not attempt to validate the URL.

see #47544