Reduce queries needed for CI artifacts in MR widget
While the individual queries here are reasonable, these methods are only called
together: so we always call `sast_artifact`, `dependency_scanning_artifact`,
etc. in the same request, and always for the merge request widget.

That means that we can save a few queries by just getting the artifacts once,
and filtering in Ruby.