Move package-and-qa to the test phase
The problem is this: when each phase starts, it checks if all jobs in earlier
phases are complete. package-and-qa is slow. As build is the first phase, if you
trigger package-and-qa before the prepare phase is finished, the entire test
phase will be blocked until package-and-qa completes, which isn't what we want.

But equally, if build came _after_ test, then package-and-qa couldn't be started
until the tests were finished. Putting it in the same stage as the tests means
we can have our cake and eat it.