Limit sidekiq-cluster concurrency to a maximum of 50 by default
By default, sidekiq-cluster will allocate N threads for N queues specified
on the command line, which in turn will require N + 5 Redis connections.
Especially when used with the `--negate` flag, this can lead to using hundreds
of Redis connections, which can lead to performance issues.

This change also adds a `-m N` option to cap the number of threads. For each
queue group, the concurrency factor will be set to min(number of queues, N).

Closes #7374